MMIOT-723 regression_nxp: crypto: Fixing memory corruption in nxp_crypto_003
In case BIG_BUFFER_SIZE is not a multiple of the size of the vector
ciph_data_ref, when copying the data, we will write over the buffer
boundaries.

In current case BIG_BUFFER_SIZE = 133120 and ciph_data_ref has a size
of 48 so it will do 2774 copies which equals to 133,152 bytes which
is an overrun of 32 bytes.

This patch modifies the copy rountine to handle all cases.

Signed-off-by: anass CHBADA <anass.chbada@nxp.com>
Signed-off-by: Franck LENORMAND <franck.lenormand@nxp.com>